+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hackney/mtfh-cli
 
+## 1.0.2
+
+### Patch Changes
+
+- a9a45cb: Fix child processes spawning on Windows OS
+
+## 1.0.1
+
+### Patch Changes
+
+- cdea5c8: Pretty output and error when packages are not in scope
+
 ## 1.0.0
 
 ### Major Changes