+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hackney/generator-mfe
 
+## 1.0.1
+
+### Patch Changes
+
+- cdea5c8: Fix issue where package.json info was unavailble in npx or yarn create
+
 ## 1.0.0
 
 ### Major Changes